Split simd_basic to headers dedicated for FNT and RingModN
- Accelerated operations for FNT are in simd_fnt.h
- Accelerated operations for RingModN are in simd_ring.h
